Adds: sym link, img in create, edit, Vue details

+++ database/migrations/2021_06_27_171326_create_post_tag_table.php
@@ -13,17 +13,19 @@ class CreatePostTagTable extends Migration
      */
     public function up()
     {
-        Schema::create('post_tag', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('users')) {
+            Schema::create('post_tag', function (Blueprint $table) {
+                $table->id();
 
-            // FK of posts
-            $table->unsignedBigInteger('post_id');
-            $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
+                // FK of posts
+                $table->unsignedBigInteger('post_id');
+                $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
 
-            // FK of tags
-            $table->unsignedBigInteger('tag_id');
-            $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
-        });
+                // FK of tags
+                $table->unsignedBigInteger('tag_id');
+                $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
+            });
+        }
     }
 
     /**
